Refactor theme bootstrap into shared partial
