+++ resources/views/layouts/app.blade.php
@@ -27,68 +27,7 @@
     <link rel="preconnect" href="https://fonts.bunny.net">
     <link href="https://fonts.bunny.net/css?family=figtree:400,500,600&display=swap" rel="stylesheet" />
     <!-- Styles (NUR CSS) -->
-    <script>
-        (() => {
-            const root = document.documentElement;
-            const prefersDark = window.matchMedia('(prefers-color-scheme: dark)');
-
-            const applyTheme = (isDark) => {
-                const nextIsDark = Boolean(isDark);
-
-                root.classList.toggle('dark', nextIsDark);
-                root.dataset.theme = nextIsDark ? 'dark' : 'light';
-
-                return root.classList.contains('dark');
-            };
-
-            const getStoredTheme = () => {
-                try {
-                    return window.localStorage.getItem('theme');
-                } catch (error) {
-                    return null;
-                }
-            };
-
-            const followsSystemPreference = (storedTheme = getStoredTheme()) => {
-                return storedTheme !== 'dark' && storedTheme !== 'light';
-            };
-
-            const applySystemTheme = (matches = prefersDark.matches, force = false) => {
-                const storedTheme = getStoredTheme();
-
-                if (!force && !followsSystemPreference(storedTheme)) {
-                    return root.classList.contains('dark');
-                }
-
-                return applyTheme(Boolean(matches));
-            };
-
-            const applyStoredTheme = (theme = getStoredTheme()) => {
-                if (theme === 'dark') {
-                    return applyTheme(true);
-                }
-
-                if (theme === 'light') {
-                    return applyTheme(false);
-                }
-
-                return applySystemTheme(prefersDark.matches, true);
-            };
-
-            window.__omxfcPrefersDark = prefersDark;
-            window.__omxfcApplySystemTheme = applySystemTheme;
-            window.__omxfcApplyStoredTheme = applyStoredTheme;
-
-            const storedTheme = getStoredTheme();
-
-            if (storedTheme === 'dark' || storedTheme === 'light') {
-                applyTheme(storedTheme === 'dark');
-                return;
-            }
-
-            applyTheme(prefersDark.matches);
-        })();
-    </script>
+    @include('layouts.partials.theme-bootstrap')
     @vite(['resources/css/app.css'])
     @livewireStyles
 </head>
