+++ resources/views/layouts/partials/theme-bootstrap.blade.php
@@ -1,0 +1,3 @@
+<script>
+{!! file_get_contents(resource_path('js/theme/bootstrap-inline.js')) !!}
+</script>
